refactor(lakehouse-backup): give stats the internal layering copy and config use (DEL-369)
Schema definition, row mapping and the recorder were one flat public package.
Split along those three seams and marked the helpers internal, leaving the five
symbols BackupJob and the integration test actually import.
