+++ iomete-lakehouse-backup/src/main/kotlin/com/iomete/backup/stats/internal/Schema.kt
@@ -1,25 +1,10 @@
-package com.iomete.backup.stats
+package com.iomete.backup.stats.internal
 
-import com.iomete.backup.config.ApplicationConfig
-import com.iomete.backup.config.HdfsConfig
-import com.iomete.backup.config.S3Config
-import com.iomete.backup.config.StorageConfig
-import com.iomete.backup.copy.CopyJobSummary
-import com.iomete.backup.copy.CopyResult
-import com.iomete.backup.copy.CopyStats
-import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.types.DataType
 import org.apache.spark.sql.types.DataTypes
 import org.apache.spark.sql.types.Metadata
 import org.apache.spark.sql.types.StructField
 import org.apache.spark.sql.types.StructType
-import java.sql.Timestamp
-import java.time.Instant
-
-const val RUNS_TABLE = "lakehouse_backup_runs"
-const val FILE_FAILURES_TABLE = "lakehouse_backup_run_file_failures"
-
-enum class RunStatus { RUNNING, SUCCEEDED, FAILED }
 
 private fun field(
     name: String,
@@ -35,7 +20,7 @@ private val DOUBLE = DataTypes.DoubleType
 private val BOOLEAN = DataTypes.BooleanType
 private val TIMESTAMP = DataTypes.TimestampType
 
-val RUNS_SCHEMA: StructType =
+internal val RUNS_SCHEMA: StructType =
     StructType(
         arrayOf(
             field("run_id", STRING, "the run ID shown in the console", nullable = false),
@@ -104,7 +89,7 @@ val RUNS_SCHEMA: StructType =
         ),
     )
 
-val FILE_FAILURES_SCHEMA: StructType =
+internal val FILE_FAILURES_SCHEMA: StructType =
     StructType(
         arrayOf(
             field("run_id", STRING, "joins to lakehouse_backup_runs.run_id", nullable = false),
@@ -121,106 +106,12 @@ val FILE_FAILURES_SCHEMA: StructType =
         ),
     )
 
-data class RunIdentity(
-    val runId: String,
-    val jobId: String?,
-    val startedBy: String?,
-) {
-    companion object {
-        fun current(spark: SparkSession): RunIdentity =
-            RunIdentity(
-                runId = spark.conf().get("spark.app.name"),
-                jobId = env("IOMETE_EXTERNAL_JOB_ID"),
-                startedBy = env("IOMETE_JOB_STARTED_BY"),
-            )
-
-        private fun env(name: String): String? = System.getenv(name)?.takeIf { it.isNotBlank() }
-    }
-}
-
-class RunProgress {
-    var filesListed: Long? = null
-    var dirsListed: Long? = null
-    var bytesSource: Long? = null
-    var sourceListingMs: Long? = null
-    var summary: CopyJobSummary? = null
-    var copy: CopyStats? = null
-    var failures: List<CopyResult> = emptyList()
-}
-
-fun runRow(
-    identity: RunIdentity,
-    config: ApplicationConfig,
-    startedAt: Instant,
-    endedAt: Instant?,
-    status: RunStatus,
-    errorMessage: String?,
-    progress: RunProgress,
-): Map<String, Any?> {
-    val summary = progress.summary
-    val copy = progress.copy
-
-    return mapOf(
-        "run_id" to identity.runId,
-        "job_id" to identity.jobId,
-        "started_by" to identity.startedBy,
-        "source_type" to storageType(config.source),
-        "source_uri" to config.source.rootUri,
-        "target_type" to storageType(config.target),
-        "target_uri" to config.target.rootUri,
-        "status" to status.name,
-        "error_message" to errorMessage,
-        "started_at" to Timestamp.from(startedAt),
-        "ended_at" to endedAt?.let { Timestamp.from(it) },
-        "files_listed" to progress.filesListed,
-        "dirs_listed" to progress.dirsListed,
-        "files_copied" to copy?.filesCopied,
-        "files_skipped" to summary?.skippedCount?.toLong(),
-        "files_failed" to summary?.failureCount?.toLong(),
-        "dirs_created" to copy?.dirsCreated,
-        "retries_used" to copy?.retriesUsed,
-        "failures_truncated" to copy?.failuresTruncated,
-        "bytes_source" to progress.bytesSource,
-        "bytes_copied" to summary?.totalBytesCopied,
-        "bytes_skipped" to summary?.skippedBytes,
-        "source_listing_ms" to progress.sourceListingMs,
-        "target_listing_ms" to copy?.targetListingMs,
-        "planning_ms" to copy?.planningMs,
-        "copy_ms" to copy?.copyMs,
-        "dir_create_ms" to copy?.dirCreateMs,
-        "copy_task_ms" to copy?.executor?.copyTaskMs,
-        "fs_init_ms" to copy?.executor?.fsInitMs,
-        "source_read_ms" to copy?.executor?.sourceReadMs,
-        "target_write_ms" to copy?.executor?.targetWriteMs,
-        "throttle_wait_ms" to copy?.executor?.throttleWaitMs,
-        "verify_ms" to copy?.executor?.verifyMs,
-        "commit_ms" to copy?.executor?.commitMs,
-        "retry_sleep_ms" to copy?.executor?.retrySleepMs,
-        "bytes_per_task" to config.copy.bytesPerTask,
-        "files_per_task" to config.copy.filesPerTask,
-        "skip_identical" to config.copy.skipIdentical,
-        "max_bandwidth_mb_per_sec" to config.copy.maxBandwidthMbPerSec,
-        "task_count" to copy?.taskCount,
-        "largest_file_bytes" to copy?.largestFileBytes,
-    )
-}
-
-fun fileFailureRow(
-    identity: RunIdentity,
-    startedAt: Instant,
-    result: CopyResult,
-): Map<String, Any?> =
-    mapOf(
-        "run_id" to identity.runId,
-        "started_at" to Timestamp.from(startedAt),
-        "source_path" to result.sourcePath,
-        "target_path" to result.targetPath,
-        "attempts_used" to result.attemptsUsed,
-        "error" to (result.error ?: "unknown"),
-    )
-
-private fun storageType(config: StorageConfig): String =
-    when (config) {
-        is S3Config -> "s3"
-        is HdfsConfig -> "hdfs"
-    }
+internal fun createTableSql(
+    table: String,
+    schema: StructType,
+): String =
+    """
+    CREATE TABLE IF NOT EXISTS $table (${schema.toDDL()})
+    USING iceberg
+    PARTITIONED BY (days(started_at))
+    """.trimIndent()
